requirements.txt
discord.py==1.5.0
python-aiml==0.9.3
pymongo[srv]
